feat: Rework console
On uefi you lose access to the vga hardware after booting, Instead you get a linear framebuffer which draw the pixels you want.

This commit implements a basic console using a framebuffer, it also provides a generic print and println macro which draws it's output if available  and sends it over serial.